# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sambelfarm Content Lab API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  password: string;
}

export interface LoginResponse {
  token: string;
}

export interface MeResponse {
  ok: boolean;
}

export interface ClaudeRequestBody {
  system?: string;
  prompt: string;
  model?: string;
}

export interface ClaudeResponse {
  result: string;
}

export type NotionQueryBodyFilter = { [key: string]: unknown };

export type NotionQueryBodySortsItem = { [key: string]: unknown };

export interface NotionQueryBody {
  database_id: string;
  filter?: NotionQueryBodyFilter;
  sorts?: NotionQueryBodySortsItem[];
  page_size?: number;
}

export type NotionQueryResponseResultsItem = { [key: string]: unknown };

export interface NotionQueryResponse {
  results: NotionQueryResponseResultsItem[];
  has_more: boolean;
  /** @nullable */
  next_cursor?: string | null;
}

export type NotionCreatePageBodyProperties = { [key: string]: unknown };

export type NotionCreatePageBodyChildrenItem = { [key: string]: unknown };

export interface NotionCreatePageBody {
  database_id: string;
  properties: NotionCreatePageBodyProperties;
  children?: NotionCreatePageBodyChildrenItem[];
}

export type NotionUpdatePageBodyProperties = { [key: string]: unknown };

export type NotionUpdatePageBodyChildrenItem = { [key: string]: unknown };

export interface NotionUpdatePageBody {
  properties: NotionUpdatePageBodyProperties;
  children?: NotionUpdatePageBodyChildrenItem[];
}

export type NotionPageResponseProperties = { [key: string]: unknown };

export interface NotionPageResponse {
  id: string;
  url?: string;
  properties?: NotionPageResponseProperties;
}