# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sambelfarm Content Lab API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with admin password
 */
export const LoginBody = zod.object({
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Validate token
 */
export const GetMeResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Claude AI proxy for script generation and TRIBE analysis
 */
export const ClaudeProxyBody = zod.object({
  system: zod.string().optional(),
  prompt: zod.string(),
  model: zod.string().optional(),
});

export const ClaudeProxyResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Query Notion database
 */
export const NotionQueryBody = zod.object({
  database_id: zod.string(),
  filter: zod.object({}).passthrough().optional(),
  sorts: zod.array(zod.object({}).passthrough()).optional(),
  page_size: zod.number().optional(),
});

export const NotionQueryResponse = zod.object({
  results: zod.array(zod.object({}).passthrough()),
  has_more: zod.boolean(),
  next_cursor: zod.string().nullish(),
});

/**
 * @summary Create a new Notion page (save script)
 */
export const NotionCreatePageBody = zod.object({
  database_id: zod.string(),
  properties: zod.object({}).passthrough(),
  children: zod.array(zod.object({}).passthrough()).optional(),
});

export const NotionCreatePageResponse = zod.object({
  id: zod.string(),
  url: zod.string().optional(),
  properties: zod.object({}).passthrough().optional(),
});

/**
 * @summary Update an existing Notion page (edit script)
 */
export const NotionUpdatePageParams = zod.object({
  pageId: zod.coerce.string(),
});

export const NotionUpdatePageBody = zod.object({
  properties: zod.object({}).passthrough(),
  children: zod.array(zod.object({}).passthrough()).optional(),
});

export const NotionUpdatePageResponse = zod.object({
  id: zod.string(),
  url: zod.string().optional(),
  properties: zod.object({}).passthrough().optional(),
});